serve: Outsource header and navigation bar
